requirements.txt
torchvision
easydict
visdom
matplotlib
Cython
pyyaml
